requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Werkzeug
python-docx